pyproject.toml
[project]
name = "fullerenefactory"
version = "0.1.0"
description = "Fullerene Factory: A Python library for generating fullerene isomers after adding addends at various positions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "rdkit>=2025.3.6",
    "crewai>=0.175.0",
    "crewai-tools>=0.69.0",
    "torch>=2.6.0",
    "huggingface_hub>=0.34.4",
    "fairchem-core>=2.4.0",
    "pubchempy>=1.0.5",
    "datasets==4.0.0",
]
